Implement possibililty for storage deployment by lookup in Enforcer and rendering via Jinja2
This allows to register scripts with Jinja2 macros. There is only one predetermined context variable for now, the preferred_storage_location, which gets preloaded by the location of the preferred storage class (if available on the named infrastructure). Effectively allowing to implement a copy-to-fast-storage as part of the run script.